refactor(stories): move Items pages to Dev Witness/Reference/Items
