+++ src/dev-witness/reference/items/items-grid.stories.tsx
@@ -2,9 +2,10 @@ import type { Meta, StoryObj } from '@storybook/react-vite';
 import { expect, within } from 'storybook/test';
 import { http, HttpResponse } from 'msw';
 import ItemsPage from '@frontend/app/items/page';
+import '@/styles/extras/vendored-theme.css';
 
 const meta: Meta<typeof ItemsPage> = {
-  title: 'Dev Witness/Items/Items Grid',
+  title: 'Dev Witness/Reference/Items/Items Grid',
   component: ItemsPage,
   tags: ['app-route:/items'],
   parameters: {
